feat(electric): Add server-side validation of null values incoming from Satellite clients  (#327)
Admitedly, we're relying on client-provided SatRelation messages to
decide whether any given column is nullable or not. So it's not a good
protection against malicious clients.

I wish we didn't have to pass nullability info via SatRelation message
from the server to the client and back. But apparently that's what we
have to do since we're not storing the history of schema changes for
every individual client.