feat: enable compensations (#340)
Compensations are stored as a special-form updates that don't specify
previous row values. Previously, such updates were treated as inserts by
the server, but now they are converted to an update that doesn't mark
any columns as modified. This allows PG to just use the tag for reviving
the row as needed.

This is covered by an e2e test.

Closes VAX-801