chore(electric): Clean up Logger calls in server-side auth logic (#402)
This addresses the issue reported in Discord on 4 Aug 2023. The
`metadata` key was a bad idea because the default console formatter for
Logger expects metadata values to be implement the `String.Chars`
protocol which lists and maps do not.

Regardless, I now think that including the error in the logged message
is better DX than tagging it on as metadata.